Add page for postcode search results
This *should* call out to the opendatacommunities API to
get real address data.

But because of ZScaler on my machine I can't actually make
the call to the API locally to test this functionality.

And actually now I've added a new key "certificate-type" to
the data maybe we should just use the local toy data and
do away with the API call altogether...

Anyway, this works OK for the toy data.