Treat long searches as a reference and go direct to certificate
The search box says "Postcode or reference" but up until now
everything took the user to the postcode results page.

This updates things so that anything longer than 7 characters
(after spaces are stripped out) is treated as a reference number
and takes the user directly to a random certificate page.